update code for simplicity
