+++ models/cnsn.py
@@ -149,6 +149,38 @@ class SelfNorm(nn.Module):
         else:
             return x * g_y.expand_as(x)
 
+class SRMLayer(nn.Module):
+    def __init__(self, channel):
+        # Reduction for compatibility with layer_block interface
+        super(SRMLayer, self).__init__()
+
+        # CFC: channel-wise fully connected layer
+        self.cfc = nn.Conv1d(channel, channel, kernel_size=2, bias=False,
+                             groups=channel)
+        self.bn = nn.BatchNorm1d(channel)
+
+    def forward(self, x): 
+        # print('srm forward...')
+        b, c, _, _ = x.size()
+
+        # Style pooling
+        mean, std = calc_ins_mean_std(x, eps=1e-8)
+        mean = mean.squeeze(3)
+        std = std.squeeze(3)
+
+        # mean = x.view(b, c, -1).mean(-1).unsqueeze(-1)
+        # std = x.view(b, c, -1).std(-1).unsqueeze(-1)
+
+        u = torch.cat((mean, std), -1)  # (b, c, 2)
+
+        # Style integration
+        z = self.cfc(u)  # (b, c, 1)
+        z = self.bn(z)
+        g = torch.sigmoid(z)
+        g = g.view(b, c, 1, 1)
+
+        return x * g.expand_as(x)
+
 
 class CNSN(nn.Module):
     """A module to combine CrossNorm and SelfNorm"""
