+++ segmentation/model/cnsn.py
@@ -1,0 +1,1 @@
+../../models/cnsn.py
